Do not try to register PZP with itself
This makes registerSender throw an exception if its from and to params
are the same, as it is not useful to try and register with itself. This
change makes it possible to bypass the following bug.

PZP code assumes the TLS port to connect with other PZPs is always the
same. Thus when using two PZPs on one host, these two cannot directly
connect as the port will be taken by the first. And the first PZP will
connect to itself.

Jira-issue: WP-250